refactor(auth): move getAuthToken function to a separate utility file for better organization
